fix warnings

+++ Team3Project1/RacerGame.cpp
@@ -146,7 +146,7 @@ void RacerGame::LoadContent() {
 	/*	road =new Road (5);
 		road->SetPhysics(5);*/
 
-	unsigned int size = 5;
+	float size = 5;
 
 	VehiclePhysicsNode* vpn = new VehiclePhysicsNode(size);
 	
@@ -220,7 +220,7 @@ void RacerGame::LoadContent() {
 	f=0;
 	b=0;
 	temp2=T3Vector3(0,0,0);
-	Speed_Rotate = 0.4;
+	Speed_Rotate = 0.4f;
 	tempPosition = T3Vector3(0,350,-800);
 	PlayerPosition=T3Vector3(500,100,-800);
